backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "apscheduler>=3.11.2",
    "beautifulsoup4>=4.14.3",
    "fastapi>=0.133.1",
    "httpx>=0.28.1",
    "pydantic>=2.12.5",
    "python-dateutil>=2.9.0.post0",
    "sqlalchemy>=2.0.47",
    "uvicorn>=0.41.0",
]
